refactor: extract model and config helpers to utils
