+++ ext/dapr-agents-ext-drasi/dapr_agents/ext/drasi/activations.py
@@ -21,13 +21,13 @@ from dapr_agents.agents.durable import DurableAgent
 from dapr_agents.agents.schemas import TriggerAction
 from dapr_agents.types.activation import ActivationContext
 from dapr_agents.types.workflow import PubSubRouteSpec
-from dapr_agents.workflow.utils.core import is_supported_model
 from dapr_agents.workflow.utils.registration import register_message_routes
 from dapr_agents.workflow.utils.subscription import (
     MessageContext,
     ModelFilter,
     TTLDedupeBackend,
 )
+from dapr_agents.utils.models import is_supported_model
 
 from dapr_agents.ext.drasi.types import DrasiOperation, DrasiChangeEvent
 from dapr_agents.ext.drasi.utils.validation import (
